Make some xpath lookups simpler, more robust.
These changes will make us more resilient to page layout changes while
also being waaay easier to visibly parse. It'd be nice to apply the same
changes to the case details page parsing, but they're a lot more dense
and tabular, which makes this challenging. There'd be value in cleaning
that up in the future.