Update SECURITY.md
This document has not been updated in a long time and given it may be used to contact us in the event of a security issue since we don't have a place on our site/we may be moving some of the content/domains, I thought it would be good to clean this up. Also I will be looking at other parts of our GH to see if we can clean up/update in other ways.